Create okhttp3 dir and move interceptor
